pyproject.toml
[project]
name = "expresso-calib"
version = "0.1.0"
description = "Live ChaRuCo intrinsic calibration station for robot camera development."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "numpy>=1.26",
  "opencv-contrib-python>=4.8",
  "pillow>=10",
  "python-multipart>=0.0.9",
  "qrcode>=7.4",
  "uvicorn[standard]>=0.30",
]

[project.scripts]
expresso-calib = "expresso_calib.server:main"

[dependency-groups]
dev = [
  "pytest>=8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/expresso_calib"]
